refactor dependencies of ext manager
