+++ pkg/cmd/extension/command.go
@@ -3,7 +3,6 @@ package extension
 import (
 	"errors"
 	"fmt"
-	"net/http"
 	"os"
 	"strings"
 	"time"
@@ -113,12 +112,7 @@ func NewCmdExtension(f *cmdutil.Factory) *cobra.Command {
 				}
 				client = api.NewCachedClient(client, time.Second*30)
 
-				cfg, err := f.Config()
-				if err != nil {
-					return err
-				}
-
-				return m.Install(client, repo, io, cfg)
+				return m.Install(repo)
 			},
 		},
 		func() *cobra.Command {
@@ -228,83 +222,9 @@ func checkValidExtension(rootCmd *cobra.Command, m extensions.ExtensionManager, 
 	return nil
 }
 
-func isBinExtension(client *http.Client, repo ghrepo.Interface) (isBin bool, err error) {
-	var r *release
-	r, err = fetchLatestRelease(client, repo)
-	if err != nil {
-		httpErr, ok := err.(api.HTTPError)
-		if ok && httpErr.StatusCode == 404 {
-			err = nil
-			return
-		}
-		return
-	}
-
-	for _, a := range r.Assets {
-		dists := possibleDists()
-		for _, d := range dists {
-			if strings.HasSuffix(a.Name, d) {
-				isBin = true
-				break
-			}
-		}
-	}
-
-	return
-}
-
 func normalizeExtensionSelector(n string) string {
 	if idx := strings.IndexRune(n, '/'); idx >= 0 {
 		n = n[idx+1:]
 	}
 	return strings.TrimPrefix(n, "gh-")
-}
-
-func possibleDists() []string {
-	return []string{
-		"aix-ppc64",
-		"android-386",
-		"android-amd64",
-		"android-arm",
-		"android-arm64",
-		"darwin-amd64",
-		"darwin-arm64",
-		"dragonfly-amd64",
-		"freebsd-386",
-		"freebsd-amd64",
-		"freebsd-arm",
-		"freebsd-arm64",
-		"illumos-amd64",
-		"ios-amd64",
-		"ios-arm64",
-		"js-wasm",
-		"linux-386",
-		"linux-amd64",
-		"linux-arm",
-		"linux-arm64",
-		"linux-mips",
-		"linux-mips64",
-		"linux-mips64le",
-		"linux-mipsle",
-		"linux-ppc64",
-		"linux-ppc64le",
-		"linux-riscv64",
-		"linux-s390x",
-		"netbsd-386",
-		"netbsd-amd64",
-		"netbsd-arm",
-		"netbsd-arm64",
-		"openbsd-386",
-		"openbsd-amd64",
-		"openbsd-arm",
-		"openbsd-arm64",
-		"openbsd-mips64",
-		"plan9-386",
-		"plan9-amd64",
-		"plan9-arm",
-		"solaris-amd64",
-		"windows-386",
-		"windows-amd64",
-		"windows-arm",
-	}
 }
