+++ pkg/cmd/extension/manager_test.go
@@ -34,7 +34,7 @@ func TestHelperProcess(t *testing.T) {
 	os.Exit(0)
 }
 
-func newTestManager(dir string) *Manager {
+func newTestManager(dir string, client *http.Client, io *iostreams.IOStreams) *Manager {
 	return &Manager{
 		dataDir:  func() string { return dir },
 		lookPath: func(exe string) (string, error) { return exe, nil },
@@ -45,6 +45,9 @@ func newTestManager(dir string) *Manager {
 			cmd.Env = []string{"GH_WANT_HELPER_PROCESS=1"}
 			return cmd
 		},
+		config: config.NewBlankConfig(),
+		io:     io,
+		client: client,
 		platform: func() string {
 			return "windows-amd64"
 		},
@@ -56,7 +59,7 @@ func TestManager_List(t *testing.T) {
 	assert.NoError(t, stubExtension(filepath.Join(tempDir, "extensions", "gh-hello", "gh-hello")))
 	assert.NoError(t, stubExtension(filepath.Join(tempDir, "extensions", "gh-two", "gh-two")))
 
-	m := newTestManager(tempDir)
+	m := newTestManager(tempDir, nil, nil)
 	exts := m.List(false)
 	assert.Equal(t, 2, len(exts))
 	assert.Equal(t, "hello", exts[0].Name())
@@ -68,7 +71,7 @@ func TestManager_Dispatch(t *testing.T) {
 	extPath := filepath.Join(tempDir, "extensions", "gh-hello", "gh-hello")
 	assert.NoError(t, stubExtension(extPath))
 
-	m := newTestManager(tempDir)
+	m := newTestManager(tempDir, nil, nil)
 
 	stdout := &bytes.Buffer{}
 	stderr := &bytes.Buffer{}
@@ -89,7 +92,7 @@ func TestManager_Remove(t *testing.T) {
 	assert.NoError(t, stubExtension(filepath.Join(tempDir, "extensions", "gh-hello", "gh-hello")))
 	assert.NoError(t, stubExtension(filepath.Join(tempDir, "extensions", "gh-two", "gh-two")))
 
-	m := newTestManager(tempDir)
+	m := newTestManager(tempDir, nil, nil)
 	err := m.Remove("hello")
 	assert.NoError(t, err)
 
@@ -105,7 +108,7 @@ func TestManager_Upgrade_AllExtensions(t *testing.T) {
 	assert.NoError(t, stubExtension(filepath.Join(tempDir, "extensions", "gh-two", "gh-two")))
 	assert.NoError(t, stubLocalExtension(tempDir, filepath.Join(tempDir, "extensions", "gh-local", "gh-local")))
 
-	m := newTestManager(tempDir)
+	m := newTestManager(tempDir, nil, nil)
 
 	stdout := &bytes.Buffer{}
 	stderr := &bytes.Buffer{}
@@ -130,7 +133,7 @@ func TestManager_Upgrade_RemoteExtension(t *testing.T) {
 	tempDir := t.TempDir()
 	assert.NoError(t, stubExtension(filepath.Join(tempDir, "extensions", "gh-remote", "gh-remote")))
 
-	m := newTestManager(tempDir)
+	m := newTestManager(tempDir, nil, nil)
 
 	stdout := &bytes.Buffer{}
 	stderr := &bytes.Buffer{}
@@ -150,7 +153,7 @@ func TestManager_Upgrade_LocalExtension(t *testing.T) {
 	tempDir := t.TempDir()
 	assert.NoError(t, stubLocalExtension(tempDir, filepath.Join(tempDir, "extensions", "gh-local", "gh-local")))
 
-	m := newTestManager(tempDir)
+	m := newTestManager(tempDir, nil, nil)
 
 	stdout := &bytes.Buffer{}
 	stderr := &bytes.Buffer{}
@@ -167,7 +170,7 @@ func TestManager_Upgrade_Force(t *testing.T) {
 
 	assert.NoError(t, stubExtension(filepath.Join(tempDir, "extensions", "gh-remote", "gh-remote")))
 
-	m := newTestManager(tempDir)
+	m := newTestManager(tempDir, nil, nil)
 
 	stdout := &bytes.Buffer{}
 	stderr := &bytes.Buffer{}
@@ -189,7 +192,7 @@ func TestManager_Upgrade_Force(t *testing.T) {
 func TestManager_Upgrade_NoExtensions(t *testing.T) {
 	tempDir := t.TempDir()
 
-	m := newTestManager(tempDir)
+	m := newTestManager(tempDir, nil, nil)
 
 	stdout := &bytes.Buffer{}
 	stderr := &bytes.Buffer{}
@@ -201,11 +204,14 @@ func TestManager_Upgrade_NoExtensions(t *testing.T) {
 
 func TestManager_Install_git(t *testing.T) {
 	tempDir := t.TempDir()
-	m := newTestManager(tempDir)
 
 	reg := httpmock.Registry{}
 	defer reg.Verify(t)
 	client := http.Client{Transport: &reg}
+
+	io, _, stdout, stderr := iostreams.Test()
+
+	m := newTestManager(tempDir, &client, io)
 
 	reg.Register(
 		httpmock.REST("GET", "repos/owner/gh-some-ext/releases/latest"),
@@ -222,11 +228,9 @@ func TestManager_Install_git(t *testing.T) {
 		httpmock.REST("GET", "repos/owner/gh-some-ext/contents/gh-some-ext"),
 		httpmock.StringResponse("script"))
 
-	io, _, stdout, stderr := iostreams.Test()
-
 	repo := ghrepo.New("owner", "gh-some-ext")
 
-	err := m.Install(&client, repo, io, config.NewBlankConfig())
+	err := m.Install(repo)
 	assert.NoError(t, err)
 	assert.Equal(t, fmt.Sprintf("[git clone https://github.com/owner/gh-some-ext.git %s]\n", filepath.Join(tempDir, "extensions", "gh-some-ext")), stdout.String())
 	assert.Equal(t, "", stderr.String())
@@ -263,12 +267,12 @@ func TestManager_Install_binary_unsupported(t *testing.T) {
 				},
 			}))
 
-	tempDir := t.TempDir()
-	m := newTestManager(tempDir)
-
 	io, _, _, _ := iostreams.Test()
+	tempDir := t.TempDir()
 
-	err := m.Install(&client, repo, io, config.NewBlankConfig())
+	m := newTestManager(tempDir, &client, io)
+
+	err := m.Install(repo)
 	assert.Error(t, err)
 
 	errText := "gh-bin-ext unsupported for windows-amd64. Open an issue: `gh issue create -R owner/gh-bin-ext -t'Support windows-amd64'`"
@@ -310,12 +314,12 @@ func TestManager_Install_binary(t *testing.T) {
 		httpmock.REST("GET", "release/cool"),
 		httpmock.StringResponse("FAKE BINARY"))
 
-	tempDir := t.TempDir()
-	m := newTestManager(tempDir)
-
 	io, _, _, _ := iostreams.Test()
+	tempDir := t.TempDir()
 
-	err := m.Install(&client, repo, io, config.NewBlankConfig())
+	m := newTestManager(tempDir, &client, io)
+
+	err := m.Install(repo)
 	assert.NoError(t, err)
 
 	manifest, err := os.ReadFile(filepath.Join(tempDir, "extensions/gh-bin-ext/manifest.yml"))
@@ -344,7 +348,7 @@ func TestManager_Create(t *testing.T) {
 	oldWd, _ := os.Getwd()
 	assert.NoError(t, os.Chdir(tempDir))
 	t.Cleanup(func() { _ = os.Chdir(oldWd) })
-	m := newTestManager(tempDir)
+	m := newTestManager(tempDir, nil, nil)
 	err := m.Create("gh-test")
 	assert.NoError(t, err)
 	files, err := ioutil.ReadDir(filepath.Join(tempDir, "gh-test"))
