+++ pkg/cmd/factory/default.go
@@ -22,7 +22,6 @@ func New(appVersion string) *cmdutil.Factory {
 		Branch:     branchFunc(), // No factory dependencies
 		Executable: executable(), // No factory dependencies
 
-		ExtensionManager: extension.NewManager(),
 	}
 
 	f.IOStreams = ioStreams(f)                   // Depends on Config
@@ -30,6 +29,7 @@ func New(appVersion string) *cmdutil.Factory {
 	f.Remotes = remotesFunc(f)                   // Depends on Config
 	f.BaseRepo = BaseRepoFunc(f)                 // Depends on Remotes
 	f.Browser = browser(f)                       // Depends on Config, and IOStreams
+	f.ExtensionManager = extensionManager(f)     // Depends on Config, HttpClient, and IOStreams
 
 	return f
 }
@@ -146,6 +146,24 @@ func branchFunc() func() (string, error) {
 		}
 		return currentBranch, nil
 	}
+}
+
+func extensionManager(f *cmdutil.Factory) *extension.Manager {
+	em := extension.NewManager(f.IOStreams)
+
+	cfg, err := f.Config()
+	if err != nil {
+		return em
+	}
+	em.SetConfig(cfg)
+
+	client, err := f.HttpClient()
+	if err != nil {
+		return em
+	}
+	em.SetClient(client)
+
+	return em
 }
 
 func ioStreams(f *cmdutil.Factory) *iostreams.IOStreams {
