+++ pkg/extensions/extension.go
@@ -2,11 +2,8 @@ package extensions
 
 import (
 	"io"
-	"net/http"
 
-	"github.com/cli/cli/v2/internal/config"
 	"github.com/cli/cli/v2/internal/ghrepo"
-	"github.com/cli/cli/v2/pkg/iostreams"
 )
 
 //go:generate moq -rm -out extension_mock.go . Extension
@@ -21,7 +18,7 @@ type Extension interface {
 //go:generate moq -rm -out manager_mock.go . ExtensionManager
 type ExtensionManager interface {
 	List(includeMetadata bool) []Extension
-	Install(*http.Client, ghrepo.Interface, *iostreams.IOStreams, config.Config) error
+	Install(ghrepo.Interface) error
 	InstallLocal(dir string) error
 	Upgrade(name string, force bool, stdout, stderr io.Writer) error
 	Remove(name string) error
